[CHORE] Invert dependency between daft-scan and daft-logical-plan
