+++ daft/daft/__init__.pyi
@@ -787,6 +787,8 @@ class ScanOperatorHandle:
     @staticmethod
     def from_python_scan_operator(operator: ScanOperator) -> ScanOperatorHandle: ...
 
+def logical_plan_table_scan(scan_operator: ScanOperatorHandle) -> LogicalPlanBuilder: ...
+
 class PartitionField:
     """
     Partitioning Field of a Scan Source such as Hive or Iceberg
@@ -1695,8 +1697,6 @@ class LogicalPlanBuilder:
         size_bytes: int,
         num_rows: int,
     ) -> LogicalPlanBuilder: ...
-    @staticmethod
-    def table_scan(scan_operator: ScanOperatorHandle) -> LogicalPlanBuilder: ...
     def with_planning_config(self, daft_planning_config: PyDaftPlanningConfig) -> LogicalPlanBuilder: ...
     def select(self, to_select: list[PyExpr]) -> LogicalPlanBuilder: ...
     def with_columns(self, columns: list[PyExpr]) -> LogicalPlanBuilder: ...
