+++ src/common/scan-info/Cargo.toml
@@ -1,0 +1,18 @@
+[dependencies]
+common-display = {path = "../display", default-features = false}
+common-error = {path = "../error", default-features = false}
+daft-dsl = {path = "../../daft-dsl", default-features = false}
+daft-schema = {path = "../../daft-schema", default-features = false}
+pyo3 = {workspace = true, optional = true}
+serde = {workspace = true}
+
+[features]
+python = ["dep:pyo3", "common-display/python", "common-error/python", "daft-dsl/python", "daft-schema/python"]
+
+[lints]
+workspace = true
+
+[package]
+edition = {workspace = true}
+name = "common-scan-info"
+version = {workspace = true}
