+++ src/daft-local-plan/Cargo.toml
@@ -1,6 +1,7 @@
 [dependencies]
 common-error = {path = "../common/error", default-features = false}
 common-resource-request = {path = "../common/resource-request", default-features = false}
+common-scan-info = {path = "../common/scan-info", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
 daft-logical-plan = {path = "../daft-logical-plan", default-features = false}
@@ -9,7 +10,7 @@ log = {workspace = true}
 strum = {version = "0.26", features = ["derive"]}
 
 [features]
-python = ["common-error/python", "common-resource-request/python", "daft-core/python", "daft-dsl/python", "daft-logical-plan/python", "daft-scan/python"]
+python = ["common-error/python", "common-resource-request/python", "common-scan-info/python", "daft-core/python", "daft-dsl/python", "daft-logical-plan/python", "daft-scan/python"]
 
 [lints]
 workspace = true
