+++ src/daft-logical-plan/Cargo.toml
@@ -20,11 +20,11 @@ common-file-formats = {path = "../common/file-formats", default-features = false
 common-io-config = {path = "../common/io-config", default-features = false}
 common-py-serde = {path = "../common/py-serde", default-features = false}
 common-resource-request = {path = "../common/resource-request", default-features = false}
+common-scan-info = {path = "../common/scan-info", default-features = false}
 common-treenode = {path = "../common/treenode", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
 daft-functions = {path = "../daft-functions", default-features = false}
-daft-scan = {path = "../daft-scan", default-features = false}
 daft-schema = {path = "../daft-schema", default-features = false}
 daft-table = {path = "../daft-table", default-features = false}
 derivative = {workspace = true}
@@ -51,11 +51,11 @@ python = [
   "common-io-config/python",
   "common-daft-config/python",
   "common-resource-request/python",
+  "common-scan-info/python",
   "daft-core/python",
   "daft-dsl/python",
   "daft-functions/python",
   "daft-table/python",
-  "daft-scan/python",
   "daft-schema/python"
 ]
