+++ src/daft-logical-plan/src/lib.rs
@@ -14,7 +14,7 @@ pub mod source_info;
 mod test;
 mod treenode;
 
-pub use builder::{LogicalPlanBuilder, ParquetScanBuilder, PyLogicalPlanBuilder};
+pub use builder::{LogicalPlanBuilder, PyLogicalPlanBuilder};
 pub use daft_core::join::{JoinStrategy, JoinType};
 pub use logical_plan::{LogicalPlan, LogicalPlanRef};
 pub use partitioning::ClusteringSpec;
@@ -30,7 +30,7 @@ use {
         python::PyFileFormatConfig, CsvSourceConfig, DatabaseSourceConfig, JsonSourceConfig,
         ParquetSourceConfig,
     },
-    daft_scan::storage_config::{NativeStorageConfig, PyStorageConfig, PythonStorageConfig},
+    // daft_scan::storage_config::{NativeStorageConfig, PyStorageConfig, PythonStorageConfig},
 };
 
 #[cfg(feature = "python")]
@@ -43,9 +43,9 @@ pub fn register_modules(parent: &Bound<PyModule>) -> PyResult<()> {
     parent.add_class::<DatabaseSourceConfig>()?;
     parent.add_class::<FileInfos>()?;
     parent.add_class::<FileInfo>()?;
-    parent.add_class::<PyStorageConfig>()?;
-    parent.add_class::<NativeStorageConfig>()?;
-    parent.add_class::<PythonStorageConfig>()?;
+    // parent.add_class::<PyStorageConfig>()?;
+    // parent.add_class::<NativeStorageConfig>()?;
+    // parent.add_class::<PythonStorageConfig>()?;
 
     Ok(())
 }
