+++ src/daft-logical-plan/src/optimization/rules/push_down_limit.rs
@@ -122,9 +122,9 @@ mod tests {
     use std::sync::Arc;
 
     use common_error::DaftResult;
+    use common_scan_info::Pushdowns;
     use daft_core::prelude::*;
     use daft_dsl::col;
-    use daft_scan::Pushdowns;
     #[cfg(feature = "python")]
     use pyo3::Python;
     use rstest::rstest;
