+++ src/daft-logical-plan/src/optimization/rules/push_down_projection.rs
@@ -658,13 +658,13 @@ mod tests {
     use std::sync::Arc;
 
     use common_error::DaftResult;
+    use common_scan_info::Pushdowns;
     use daft_core::prelude::*;
     use daft_dsl::{
         col,
         functions::python::{RuntimePyObject, UDFRuntimeBinding},
         lit,
     };
-    use daft_scan::Pushdowns;
 
     use crate::{
         optimization::{rules::PushDownProjection, test::assert_optimized_plan_with_rules_eq},
