+++ src/daft-micropartition/Cargo.toml
@@ -4,6 +4,7 @@ bincode = {workspace = true}
 common-error = {path = "../common/error", default-features = false}
 common-file-formats = {path = "../common/file-formats", default-features = false}
 common-runtime = {path = "../common/runtime", default-features = false}
+common-scan-info = {path = "../common/scan-info", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-csv = {path = "../daft-csv", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
@@ -18,7 +19,7 @@ pyo3 = {workspace = true, optional = true}
 snafu = {workspace = true}
 
 [features]
-python = ["dep:pyo3", "common-error/python", "common-file-formats/python", "daft-core/python", "daft-dsl/python", "daft-table/python", "daft-io/python", "daft-parquet/python", "daft-scan/python", "daft-stats/python"]
+python = ["dep:pyo3", "common-error/python", "common-file-formats/python", "common-scan-info/python", "daft-core/python", "daft-dsl/python", "daft-table/python", "daft-io/python", "daft-parquet/python", "daft-scan/python", "daft-stats/python"]
 
 [lints]
 workspace = true
