+++ src/daft-physical-plan/Cargo.toml
@@ -4,6 +4,7 @@ common-display = {path = "../common/display", default-features = false}
 common-error = {path = "../common/error", default-features = false}
 common-file-formats = {path = "../common/file-formats", default-features = false}
 common-resource-request = {path = "../common/resource-request", default-features = false}
+common-scan-info = {path = "../common/scan-info", default-features = false}
 common-treenode = {path = "../common/treenode", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
@@ -19,7 +20,7 @@ serde = {workspace = true, features = ["rc"]}
 rstest = {workspace = true}
 
 [features]
-python = ["common-error/python", "common-resource-request/python", "daft-core/python", "daft-dsl/python", "daft-logical-plan/python", "daft-scan/python"]
+python = ["common-error/python", "common-resource-request/python", "common-scan-info/python", "daft-core/python", "daft-dsl/python", "daft-logical-plan/python", "daft-scan/python"]
 
 [lints]
 workspace = true
