+++ src/daft-scan/src/lib.rs
@@ -1,20 +1,12 @@
 #![feature(if_let_guard)]
 #![feature(let_chains)]
-use std::{
-    borrow::Cow,
-    fmt::{Debug, Display},
-    hash::{Hash, Hasher},
-    sync::Arc,
-};
+use std::{any::Any, borrow::Cow, fmt::Debug, sync::Arc};
 
 use common_display::DisplayAs;
-use common_error::{DaftError, DaftResult};
+use common_error::DaftError;
 use common_file_formats::FileFormatConfig;
-use daft_dsl::ExprRef;
-use daft_schema::{
-    field::Field,
-    schema::{Schema, SchemaRef},
-};
+use common_scan_info::{Pushdowns, ScanTaskLike};
+use daft_schema::schema::{Schema, SchemaRef};
 use daft_stats::{PartitionSpec, TableMetadata, TableStatistics};
 use itertools::Itertools;
 use parquet2::metadata::FileMetaData;
@@ -25,6 +17,7 @@ pub use anonymous::AnonymousScanOperator;
 pub mod glob;
 mod hive;
 use common_daft_config::DaftExecutionConfig;
+pub mod builder;
 pub mod scan_task_iters;
 
 #[cfg(feature = "python")]
@@ -36,8 +29,6 @@ use pyo3::PyErr;
 pub use python::register_modules;
 use snafu::Snafu;
 use storage_config::StorageConfig;
-mod expr_rewriter;
-pub use expr_rewriter::{rewrite_predicate_for_partitioning, PredicateGroups};
 #[derive(Debug, Snafu)]
 pub enum Error {
     #[cfg(feature = "python")]
@@ -379,6 +370,23 @@ pub struct ScanTask {
     pub statistics: Option<TableStatistics>,
     pub generated_fields: Option<SchemaRef>,
 }
+
+impl ScanTaskLike for ScanTask {
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+
+    fn as_any_arc(self: Arc<Self>) -> Arc<dyn Any + Send + Sync> {
+        self
+    }
+}
+
+impl From<ScanTask> for Arc<dyn ScanTaskLike> {
+    fn from(task: ScanTask) -> Self {
+        Arc::new(task)
+    }
+}
+
 pub type ScanTaskRef = Arc<ScanTask>;
 
 impl ScanTask {
@@ -711,328 +719,6 @@ Pushdowns = {pushdowns}
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
-pub struct PartitionField {
-    field: Field,
-    source_field: Option<Field>,
-    transform: Option<PartitionTransform>,
-}
-
-impl PartitionField {
-    pub fn new(
-        field: Field,
-        source_field: Option<Field>,
-        transform: Option<PartitionTransform>,
-    ) -> DaftResult<Self> {
-        match (&source_field, &transform) {
-            (Some(_), Some(_)) => {
-                // TODO ADD VALIDATION OF TRANSFORM based on types
-                Ok(Self {
-                    field,
-                    source_field,
-                    transform,
-                })
-            }
-            (None, Some(tfm)) => Err(DaftError::ValueError(format!(
-                "transform set in PartitionField: {tfm} but source_field not set"
-            ))),
-            _ => Ok(Self {
-                field,
-                source_field,
-                transform,
-            }),
-        }
-    }
-
-    pub fn clone_field(&self) -> Field {
-        self.field.clone()
-    }
-}
-
-impl Display for PartitionField {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if let Some(tfm) = &self.transform {
-            write!(
-                f,
-                "PartitionField({}, src={}, tfm={})",
-                self.field,
-                self.source_field.as_ref().unwrap(),
-                tfm
-            )
-        } else {
-            write!(f, "PartitionField({})", self.field)
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Hash)]
-pub enum PartitionTransform {
-    /// https://iceberg.apache.org/spec/#partitioning
-    /// For Delta, Hudi and Hive, it should always be `Identity`.
-    Identity,
-    IcebergBucket(u64),
-    IcebergTruncate(u64),
-    Year,
-    Month,
-    Day,
-    Hour,
-    Void,
-}
-
-impl PartitionTransform {
-    #[must_use]
-    pub fn supports_equals(&self) -> bool {
-        true
-    }
-
-    #[must_use]
-    pub fn supports_not_equals(&self) -> bool {
-        matches!(self, Self::Identity)
-    }
-
-    #[must_use]
-    pub fn supports_comparison(&self) -> bool {
-        use PartitionTransform::{Day, Hour, IcebergTruncate, Identity, Month, Year};
-        matches!(
-            self,
-            Identity | IcebergTruncate(_) | Year | Month | Day | Hour
-        )
-    }
-}
-
-impl Display for PartitionTransform {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{self:?}")
-    }
-}
-
-pub trait ScanOperator: Send + Sync + Debug {
-    fn schema(&self) -> SchemaRef;
-    fn partitioning_keys(&self) -> &[PartitionField];
-    fn file_path_column(&self) -> Option<&str>;
-    // Although generated fields are often added to the partition spec, generated fields and
-    // partition fields are handled differently:
-    // 1. Generated fields: Currently from file paths or Hive partitions,
-    //    although in the future these may be extended to generated and virtual columns.
-    // 2. Partition fields: Originally from Iceberg, specifying both source and (possibly
-    //    transformed) partition fields.
-    //
-    // Partition fields are automatically included in scan output schemas (e.g.,
-    // in ScanTask::materialized_schema), while generated fields require special handling.
-    // Thus, we maintain separate representations for partitioning keys and generated fields.
-    fn generated_fields(&self) -> Option<SchemaRef>;
-
-    fn can_absorb_filter(&self) -> bool;
-    fn can_absorb_select(&self) -> bool;
-    fn can_absorb_limit(&self) -> bool;
-    fn multiline_display(&self) -> Vec<String>;
-    fn to_scan_tasks(
-        &self,
-        pushdowns: Pushdowns,
-    ) -> DaftResult<Box<dyn Iterator<Item = DaftResult<ScanTaskRef>>>>;
-}
-
-impl Display for dyn ScanOperator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.multiline_display().join("\n"))
-    }
-}
-
-/// Light transparent wrapper around an Arc<dyn ScanOperator> that implements Eq/PartialEq/Hash
-/// functionality to be performed on the **pointer** instead of on the value in the pointer.
-///
-/// This lets us get around having to implement full hashing/equality on [`ScanOperator`]`, which
-/// is difficult because we sometimes have weird Python implementations that can be hard to check.
-///
-/// [`ScanOperatorRef`] should be thus held by structs that need to check the "sameness" of the
-/// underlying ScanOperator instance, for example in the Scan nodes in a logical plan which need
-/// to check for sameness of Scan nodes during plan optimization.
-#[derive(Debug, Clone)]
-pub struct ScanOperatorRef(pub Arc<dyn ScanOperator>);
-
-impl Hash for ScanOperatorRef {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        Arc::as_ptr(&self.0).hash(state);
-    }
-}
-
-impl PartialEq<Self> for ScanOperatorRef {
-    fn eq(&self, other: &Self) -> bool {
-        Arc::ptr_eq(&self.0, &other.0)
-    }
-}
-
-impl std::cmp::Eq for ScanOperatorRef {}
-
-impl Display for ScanOperatorRef {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        std::fmt::Display::fmt(&self.0, f)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct PhysicalScanInfo {
-    pub scan_op: ScanOperatorRef,
-    pub source_schema: SchemaRef,
-    pub partitioning_keys: Vec<PartitionField>,
-    pub pushdowns: Pushdowns,
-}
-
-impl PhysicalScanInfo {
-    #[must_use]
-    pub fn new(
-        scan_op: ScanOperatorRef,
-        source_schema: SchemaRef,
-        partitioning_keys: Vec<PartitionField>,
-        pushdowns: Pushdowns,
-    ) -> Self {
-        Self {
-            scan_op,
-            source_schema,
-            partitioning_keys,
-            pushdowns,
-        }
-    }
-
-    #[must_use]
-    pub fn with_pushdowns(&self, pushdowns: Pushdowns) -> Self {
-        Self {
-            scan_op: self.scan_op.clone(),
-            source_schema: self.source_schema.clone(),
-            partitioning_keys: self.partitioning_keys.clone(),
-            pushdowns,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
-pub struct Pushdowns {
-    /// Optional filters to apply to the source data.
-    pub filters: Option<ExprRef>,
-    /// Optional filters to apply on partitioning keys.
-    pub partition_filters: Option<ExprRef>,
-    /// Optional columns to select from the source data.
-    pub columns: Option<Arc<Vec<String>>>,
-    /// Optional number of rows to read.
-    pub limit: Option<usize>,
-}
-
-impl Default for Pushdowns {
-    fn default() -> Self {
-        Self::new(None, None, None, None)
-    }
-}
-
-impl Pushdowns {
-    #[must_use]
-    pub fn new(
-        filters: Option<ExprRef>,
-        partition_filters: Option<ExprRef>,
-        columns: Option<Arc<Vec<String>>>,
-        limit: Option<usize>,
-    ) -> Self {
-        Self {
-            filters,
-            partition_filters,
-            columns,
-            limit,
-        }
-    }
-
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.filters.is_none()
-            && self.partition_filters.is_none()
-            && self.columns.is_none()
-            && self.limit.is_none()
-    }
-
-    #[must_use]
-    pub fn with_limit(&self, limit: Option<usize>) -> Self {
-        Self {
-            filters: self.filters.clone(),
-            partition_filters: self.partition_filters.clone(),
-            columns: self.columns.clone(),
-            limit,
-        }
-    }
-
-    #[must_use]
-    pub fn with_filters(&self, filters: Option<ExprRef>) -> Self {
-        Self {
-            filters,
-            partition_filters: self.partition_filters.clone(),
-            columns: self.columns.clone(),
-            limit: self.limit,
-        }
-    }
-
-    #[must_use]
-    pub fn with_partition_filters(&self, partition_filters: Option<ExprRef>) -> Self {
-        Self {
-            filters: self.filters.clone(),
-            partition_filters,
-            columns: self.columns.clone(),
-            limit: self.limit,
-        }
-    }
-
-    #[must_use]
-    pub fn with_columns(&self, columns: Option<Arc<Vec<String>>>) -> Self {
-        Self {
-            filters: self.filters.clone(),
-            partition_filters: self.partition_filters.clone(),
-            columns,
-            limit: self.limit,
-        }
-    }
-
-    #[must_use]
-    pub fn multiline_display(&self) -> Vec<String> {
-        let mut res = vec![];
-        if let Some(columns) = &self.columns {
-            res.push(format!("Projection pushdown = [{}]", columns.join(", ")));
-        }
-        if let Some(filters) = &self.filters {
-            res.push(format!("Filter pushdown = {filters}"));
-        }
-        if let Some(pfilters) = &self.partition_filters {
-            res.push(format!("Partition Filter = {pfilters}"));
-        }
-        if let Some(limit) = self.limit {
-            res.push(format!("Limit pushdown = {limit}"));
-        }
-        res
-    }
-}
-impl DisplayAs for Pushdowns {
-    fn display_as(&self, level: common_display::DisplayLevel) -> String {
-        match level {
-            common_display::DisplayLevel::Compact => {
-                let mut s = String::new();
-                s.push_str("Pushdowns: {");
-                let mut sub_items = vec![];
-                if let Some(columns) = &self.columns {
-                    sub_items.push(format!("projection: [{}]", columns.join(", ")));
-                }
-                if let Some(filters) = &self.filters {
-                    sub_items.push(format!("filter: {filters}"));
-                }
-                if let Some(pfilters) = &self.partition_filters {
-                    sub_items.push(format!("partition_filter: {pfilters}"));
-                }
-                if let Some(limit) = self.limit {
-                    sub_items.push(format!("limit: {limit}"));
-                }
-                s.push_str(&sub_items.join(", "));
-                s.push('}');
-                s
-            }
-            _ => self.multiline_display().join("\n"),
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
     use std::sync::Arc;
@@ -1040,13 +726,14 @@ mod test {
     use common_display::{DisplayAs, DisplayLevel};
     use common_error::DaftResult;
     use common_file_formats::{FileFormatConfig, ParquetSourceConfig};
+    use common_scan_info::{Pushdowns, ScanOperator};
     use daft_schema::{schema::Schema, time_unit::TimeUnit};
     use itertools::Itertools;
 
     use crate::{
         glob::GlobScanOperator,
         storage_config::{NativeStorageConfig, StorageConfig},
-        DataSource, Pushdowns, ScanOperator, ScanTask,
+        DataSource, ScanTask,
     };
 
     fn make_scan_task(num_sources: usize) -> ScanTask {
