+++ src/daft-sql/Cargo.toml
@@ -7,6 +7,7 @@ daft-dsl = {path = "../daft-dsl"}
 daft-functions = {path = "../daft-functions"}
 daft-functions-json = {path = "../daft-functions-json"}
 daft-logical-plan = {path = "../daft-logical-plan"}
+daft-scan = {path = "../daft-scan"}
 once_cell = {workspace = true}
 pyo3 = {workspace = true, optional = true}
 sqlparser = {workspace = true}
@@ -17,7 +18,7 @@ snafu.workspace = true
 rstest = {workspace = true}
 
 [features]
-python = ["dep:pyo3", "common-error/python", "daft-functions/python", "daft-functions-json/python"]
+python = ["dep:pyo3", "common-error/python", "daft-functions/python", "daft-functions-json/python", "daft-scan/python"]
 
 [lints]
 workspace = true
