+++ src/daft-sql/src/table_provider/mod.rs
@@ -98,7 +98,7 @@ impl SQLTableFunction for ReadDeltalakeFunction {
             unsupported_sql_err!("Expected a string literal for the first argument");
         };
 
-        LogicalPlanBuilder::delta_scan(uri, io_config, true).map_err(From::from)
+        daft_scan::builder::delta_scan(uri, io_config, true).map_err(From::from)
     }
 }
